requirements.txt
fastapi
uvicorn[standard]
pydantic
pillow
pytesseract
pdf2image
opencv-python-headless
numpy
requests
pytest
regex
python-dotenv
